pyproject.toml
[project]
name = "trapping-immunization"
version = "0.1.0"
description = "From-scratch reimplementation of Sarker et al. NeurIPS 2025 Lock-LLM workshop trapping paper"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "timm>=1.0",
    "datasets>=2.18",
    "Pillow",
    "numpy",
    "pyyaml",
    "tqdm",
    "wandb",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
